VS-44 Rename components
