+++ vera-art/src/components/UI/AppFooter.vue
