+++ vera-art/src/components/UI/PicCarousel.vue
