+++ vera-art/src/views/AllWorks.vue
@@ -1,5 +1,5 @@
 <script lang="ts">
-import Gallery from '@/components/UI/Gallery.vue';
+import Gallery from '@/components/UI/PicGallery.vue';
 import axios from 'axios';
 import { useI18n } from 'vue-i18n';
